+++ v3/crates/execute/src/lib.rs
@@ -6,6 +6,7 @@ pub mod model_tracking;
 pub mod ndc;
 mod plan;
 mod process_response;
+mod query_usage;
 mod remote_joins;
 
 pub use plan::process_model_relationship_definition;
@@ -68,13 +69,19 @@ impl<'a> TraceableError for GraphQLErrors<'a> {
 pub struct GraphQLResponse(gql::http::Response);
 
 impl GraphQLResponse {
-    pub fn from_result(result: ExecuteQueryResult) -> Self {
-        Self(result.to_graphql_response())
+    pub fn from_result(
+        result: ExecuteQueryResult,
+        expose_internal_errors: ExposeInternalErrors,
+    ) -> Self {
+        Self(result.to_graphql_response(expose_internal_errors))
     }
 
-    pub fn from_error(err: &error::RequestError) -> Self {
+    pub fn from_error(
+        err: &error::RequestError,
+        expose_internal_errors: ExposeInternalErrors,
+    ) -> Self {
         Self(Response::error(
-            err.to_graphql_error(),
+            err.to_graphql_error(expose_internal_errors),
             axum::http::HeaderMap::default(),
         ))
     }
@@ -105,6 +112,7 @@ impl Traceable for GraphQLResponse {
 pub struct ProjectId(pub String);
 
 pub async fn execute_query(
+    expose_internal_errors: ExposeInternalErrors,
     http_context: &HttpContext,
     schema: &Schema<GDS>,
     session: &Session,
@@ -113,6 +121,7 @@ pub async fn execute_query(
     project_id: Option<&ProjectId>,
 ) -> GraphQLResponse {
     execute_query_internal(
+        expose_internal_errors,
         http_context,
         schema,
         session,
@@ -121,7 +130,7 @@ pub async fn execute_query(
         project_id,
     )
     .await
-    .unwrap_or_else(|e| GraphQLResponse::from_error(&e))
+    .unwrap_or_else(|e| GraphQLResponse::from_error(&e, expose_internal_errors))
 }
 
 #[derive(Error, Debug)]
@@ -143,8 +152,15 @@ impl TraceableError for GraphQlValidationError {
     }
 }
 
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub enum ExposeInternalErrors {
+    Expose,
+    Censor,
+}
+
 /// Executes a GraphQL query
 pub async fn execute_query_internal(
+    expose_internal_errors: ExposeInternalErrors,
     http_context: &HttpContext,
     schema: &gql::schema::Schema<GDS>,
     session: &Session,
@@ -194,13 +210,14 @@ pub async fn execute_query_internal(
                             let all_usage_counts =
                                 model_tracking::get_all_usage_counts_in_query(&ir);
                             let serialized_data = serde_json::to_string(&all_usage_counts).unwrap();
+
                             set_attribute_on_active_span(
                                 AttributeVisibility::Default,
                                 "usage_counts",
                                 serialized_data,
                             );
 
-                            Box::pin(async {
+                            let execute_response = Box::pin(async {
                                 let execute_query_result = match request_plan {
                                     plan::RequestPlan::MutationPlan(mutation_plan) => {
                                         plan::execute_mutation_plan(
@@ -219,8 +236,35 @@ pub async fn execute_query_internal(
                                         .await
                                     }
                                 };
-                                GraphQLResponse::from_result(execute_query_result)
-                            })
+
+                                GraphQLResponse::from_result(
+                                    execute_query_result,
+                                    expose_internal_errors,
+                                )
+                            });
+
+                            // Analyze the query usage
+                            // It is attached to this span as an attribute
+                            match analyze_query_usage(&normalized_request) {
+                                Err(analyze_error) => {
+                                    // Set query usage analytics error as a span attribute
+                                    set_attribute_on_active_span(
+                                        AttributeVisibility::Internal,
+                                        "query_usage_analytics_error",
+                                        analyze_error.to_string(),
+                                    );
+                                }
+                                Ok(query_usage_analytics) => {
+                                    // Set query usage analytics as a span attribute
+                                    set_attribute_on_active_span(
+                                        AttributeVisibility::Internal,
+                                        "query_usage_analytics",
+                                        query_usage_analytics,
+                                    );
+                                }
+                            }
+
+                            execute_response
                         })
                         .await;
                     Ok(response)
@@ -232,6 +276,7 @@ pub async fn execute_query_internal(
 
 /// Explains (query plan) a GraphQL query
 pub async fn explain_query_internal(
+    expose_internal_errors: ExposeInternalErrors,
     http_context: &HttpContext,
     schema: &gql::schema::Schema<GDS>,
     session: &Session,
@@ -280,6 +325,7 @@ pub async fn explain_query_internal(
                                     let request_result = match request_plan {
                                         plan::RequestPlan::MutationPlan(mutation_plan) => {
                                             crate::explain::explain_mutation_plan(
+                                                expose_internal_errors,
                                                 http_context,
                                                 mutation_plan,
                                             )
@@ -287,6 +333,7 @@ pub async fn explain_query_internal(
                                         }
                                         plan::RequestPlan::QueryPlan(query_plan) => {
                                             crate::explain::explain_query_plan(
+                                                expose_internal_errors,
                                                 http_context,
                                                 query_plan,
                                             )
@@ -297,7 +344,7 @@ pub async fn explain_query_internal(
                                     match request_result {
                                         Ok(step) => step.make_explain_response(),
                                         Err(e) => explain::types::ExplainResponse::error(
-                                            e.to_graphql_error(),
+                                            e.to_graphql_error(expose_internal_errors),
                                         ),
                                     }
                                 })
@@ -437,6 +484,21 @@ pub fn generate_ir<'n, 's>(
     }
 }
 
+fn analyze_query_usage<'s>(
+    normalized_request: &'s Operation<'s, GDS>,
+) -> Result<String, error::QueryUsageAnalyzeError> {
+    let tracer = tracing_util::global_tracer();
+    tracer.in_span(
+        "analyze_query_usage",
+        "Analyze query usage",
+        SpanVisibility::Internal,
+        || {
+            let query_usage_analytics = query_usage::analyze_query_usage(normalized_request);
+            Ok(serde_json::to_string(&query_usage_analytics)?)
+        },
+    )
+}
+
 #[cfg(test)]
 mod tests {
     use goldenfile::{differs::text_diff, Mint};
@@ -453,6 +515,7 @@ mod tests {
     };
 
     use super::generate_ir;
+    use super::query_usage::analyze_query_usage;
     use schema::GDS;
 
     #[test]
@@ -513,6 +576,68 @@ mod tests {
             write!(expected, "{}", serde_json::to_string_pretty(&ir)?)?;
         }
 
+        Ok(())
+    }
+
+    #[test]
+    fn test_query_usage_analytics() -> Result<(), Box<dyn std::error::Error>> {
+        let test_dir = PathBuf::from(env!("CARGO_MANIFEST_DIR"))
+            .join("tests")
+            .join("query_usage_analytics");
+        let mut mint = Mint::new(&test_dir);
+
+        let schema = fs::read_to_string(test_dir.join("schema.json"))?;
+
+        let gds = GDS::new_with_default_flags(open_dds::Metadata::from_json_str(&schema)?)?;
+        let schema = GDS::build_schema(&gds)?;
+
+        for test_dir in fs::read_dir(test_dir)? {
+            let path = test_dir?.path();
+
+            if !path.is_dir() {
+                continue;
+            }
+
+            let test_name = path
+                .file_name()
+                .ok_or_else(|| format!("{path:?} is not a normal file or directory"))?;
+
+            let raw_request = fs::read_to_string(path.join("request.gql"))?;
+            let expected_path = PathBuf::from(test_name).join("expected.json");
+
+            let session_vars_path = path.join("session_variables.json");
+            let session = resolve_session(session_vars_path);
+            let query = Parser::new(&raw_request).parse_executable_document()?;
+
+            let request = Request {
+                operation_name: None,
+                query,
+                variables: HashMap::new(),
+            };
+
+            let normalized_request = normalize_request(
+                &schema::GDSRoleNamespaceGetter {
+                    scope: session.role.clone(),
+                },
+                &schema,
+                &request,
+            )?;
+
+            let query_usage = analyze_query_usage(&normalized_request);
+            let mut expected = mint.new_goldenfile_with_differ(
+                expected_path,
+                Box::new(|file1, file2| {
+                    let json1: serde_json::Value =
+                        serde_json::from_reader(File::open(file1).unwrap()).unwrap();
+                    let json2: serde_json::Value =
+                        serde_json::from_reader(File::open(file2).unwrap()).unwrap();
+                    if json1 != json2 {
+                        text_diff(file1, file2);
+                    }
+                }),
+            )?;
+            write!(expected, "{}", serde_json::to_string_pretty(&query_usage)?)?;
+        }
         Ok(())
     }