+++ v3/crates/metadata-resolve/src/stages/mod.rs
@@ -1,5 +1,6 @@
 pub mod aggregates;
 mod apollo;
+mod arguments;
 pub mod boolean_expressions;
 pub mod command_permissions;
 pub mod commands;
@@ -13,19 +14,20 @@ pub mod object_boolean_expressions;
 pub mod object_types;
 pub mod relationships;
 pub mod roles;
+pub mod scalar_boolean_expressions;
 pub mod scalar_types;
 pub mod type_permissions;
 mod types;
 
 pub use types::Metadata;
 
+use crate::types::configuration::Configuration;
 use crate::types::error::Error;
-use crate::types::internal_flags::MetadataResolveFlagsInternal;
 
 /// This is where we take the input metadata and attempt to resolve a working `Metadata` object.
 pub fn resolve(
     metadata: open_dds::Metadata,
-    flags: MetadataResolveFlagsInternal,
+    configuration: Configuration,
 ) -> Result<Metadata, Error> {
     let metadata_accessor: open_dds::accessor::MetadataAccessor =
         open_dds::accessor::MetadataAccessor::new(metadata);
@@ -36,7 +38,7 @@ pub fn resolve(
         graphql_config::resolve(&metadata_accessor.graphql_config, metadata_accessor.flags)?;
 
     // Fetch and check schema information for all our data connectors
-    let data_connectors = data_connectors::resolve(&metadata_accessor)?;
+    let data_connectors = data_connectors::resolve(&metadata_accessor, &configuration)?;
 
     // Validate object types defined in metadata
     let object_types::DataConnectorTypeMappingsOutput {
@@ -52,6 +54,17 @@ pub fn resolve(
         graphql_types,
     } = scalar_types::resolve(&metadata_accessor, &graphql_types)?;
 
+    // Validate scalar `BooleanExpressionType`s
+    let scalar_boolean_expressions::ScalarBooleanExpressionsOutput {
+        graphql_types,
+        boolean_expression_scalar_types,
+    } = scalar_boolean_expressions::resolve(
+        &metadata_accessor,
+        configuration,
+        &graphql_types,
+        &data_connectors,
+    )?;
+
     // Validate `DataConnectorScalarType` metadata. This will soon be deprecated and subsumed by
     // `BooleanExpressionType`
     let data_connector_scalar_types::DataConnectorWithScalarsOutput {
@@ -61,12 +74,26 @@ pub fn resolve(
         &metadata_accessor,
         &data_connectors,
         &scalar_types,
+        &boolean_expression_scalar_types,
         &graphql_types,
     )?;
 
     // Fetch and validate permissions, and attach them to the relevant object types
     let object_types_with_permissions =
         type_permissions::resolve(&metadata_accessor, &object_types)?;
+
+    // Resolve fancy new boolean expression types
+    let boolean_expressions::BooleanExpressionsOutput {
+        boolean_expression_types,
+        graphql_types,
+    } = boolean_expressions::resolve(
+        &metadata_accessor,
+        configuration,
+        &boolean_expression_scalar_types,
+        &graphql_types,
+        &graphql_config,
+        &object_types_with_permissions,
+    )?;
 
     // Check aggregate expressions
     let aggregates::AggregateExpressionsOutput {
@@ -94,19 +121,6 @@ pub fn resolve(
         &object_types_with_permissions,
         &graphql_types,
         &graphql_config,
-    )?;
-
-    // Resolve fancy new boolean expression types
-    let boolean_expressions::BooleanExpressionsOutput {
-        boolean_expression_types,
-        graphql_types,
-    } = boolean_expressions::resolve(
-        &metadata_accessor,
-        flags,
-        &graphql_types,
-        &graphql_config,
-        &data_connectors,
-        &object_types_with_permissions,
     )?;
 
     // Resolve models and their sources
@@ -143,6 +157,7 @@ pub fn resolve(
 
     let object_types_with_relationships = relationships::resolve(
         &metadata_accessor,
+        configuration,
         &data_connectors,
         &data_connector_scalars,
         &object_types_with_permissions,
@@ -152,12 +167,20 @@ pub fn resolve(
         &graphql_config,
     )?;
 
+    // now we know about relationships, we can check our arguments (particularly, any
+    // boolean expressions they use and whether their relationships are valid)
+    arguments::resolve(
+        &commands,
+        &models,
+        &object_types_with_relationships,
+        &scalar_types,
+        &object_boolean_expression_types,
+        &boolean_expression_types,
+    )?;
+
     // Resolve the filter expressions and graphql settings for models
     // This is a separate step so we can look up resolved models and their sources
-    let models_graphql::ModelsGraphqlOutput {
-        models_with_graphql,
-        graphql_types: _,
-    } = models_graphql::resolve(
+    let models_with_graphql = models_graphql::resolve(
         &metadata_accessor,
         &models,
         &data_connector_scalars,